Add type doc comments and isolate CTS1 related types
